Modify mo_query_builder so that it returns a GMOQuery object
It does this indirectly, by returning a function, which, when passed
a bunch of keyword/value pairs returns a GMOQuery object.

Restrict DbusClientMissingSearchPropertiesError so that it does not take
an object path argument. In the context in which it is now raised the
object path is not available. The fact that the object path was used
initially, was in the nature of an oversight/bug, so it is nice that that
is fixed.

Update tests as appropriate.

Signed-off-by: mulhern <amulhern@redhat.com>