Rewrite Hypothesis-based testing
Make all strategies with multiple sub-strategies composite.
This allows finer control over the size of the test objects, and helps
to reduce the number of long-running tests that can be generated.

Limit the size of interface names that can be generated by the strategy.

Use the new parameters added to the interface strategy in tests to reduce
the size of the objects.

Reduce the number of examples generated for each test.

Incorporate a test that causes a MissingPropertyError and remove the
"no cover" pragma on that error.

Use the tuple strategy for a test to ensure that both a spec that defines
no properties and one that defines at least one property are generated
for the test.

Comment all remaining "no cover" pragmas.

In tests, only generate very minimal example tables. Complete example
tables should be generated by a Hypothesis strategy instead. The net
effect of all this is to remove one "no cover" pragma on an error type and
to add one "no cover" pragma to a line in GMOQuery.search(), which
is not bad at all.

Signed-off-by: mulhern <amulhern@redhat.com>